requirements.txt
scipy==1.1.0
tqdm==4.23.4
numpy==1.14.3
pynput==1.3.7
torch==0.3.1
matplotlib==2.2.2
seaborn==0.8.1
opencv_python==3.3.0.9
PyAudio==0.2.11
setuptools==39.2.0
cloudpickle==0.5.3
gym==0.10.5
madmom==0.15.1
tensorflow==1.5.0
tensorboardX==1.2
PyYAML==3.12
